rfctr: move _Subtotal into dimension module
Dimension is the only client of _Subtotal. Move it into the dimension
module and remove the subtotal module.
